Rename android application module to avoid naming conflicts
Due to the actual `:taskodoro` name of the android application module
can result in gradle conflicts with the expected unified shared module
that will have that same name.
