+++ apps/android/app/.gitignore
