+++ apps/android/app/proguard-rules.pro
