+++ apps/android/app/src/main/AndroidManifest.xml
@@ -25,9 +25,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Taskodoro">
         <activity
-            android:name=".MainActivity"
+            android:name=".app.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.Taskodoro">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
