+++ apps/android/app/src/main/res/values/colors.xml
